requirements.txt
py-clob-client>=0.18.0
websockets>=12.0
httpx>=0.25.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
ccxt>=4.0.0
python-telegram-bot>=20.0
